Fixed Versioning
In the future, our current setup would cause headaches if we need to put new versions in the database for a live test. Instead of grabbing the biggest version from the database, which would result in current users being alerted to a new app version (which doesn't exist yet) while we are making it, we will instead make the current version a static script which we can change when we are ready to release the update.

This would've been done in the future if I hadn't done it already.