fix(login): Fix various mistakes and bugs
First of all, I forgot to uncomment the line in the /lib/api.js file which stops execution if the
subdomain wasn't "api". I also screwed up in making sure that the proper tokens are returned when
the prior one is outdated, and also resulted in trying to add an already existing account, and then
later crashing due to attempts to send headers twice inside of a callback. Also, the spacing for
some file headers was messed up and not correctly centered. Perfection at last.