feat(signup): Update signup to actually do what it was meant to do
This update includes changes to the api and server files. Namely, email has been added to the signup
process and templates to go with it. The server has better logging during startup, and the timer
updated to actually time to full startup process. Some minor fixes to html and a few syntax
corrections that I somehow missed in an earlier commit. Yes, those syntax errors crashed the app and
just goes to show how important actually running your code is.

full login and signup done #18