feat(signup): Finish account activation with email links
Although the previous commit included an email link system, there was no actual email link
functionality implemented. This brings that to the process in a very simple way, as well as updating
links with the time the token was created, just to further reduce the possibility of duplication.
This also sliently includes a much better response timing system which is implemented as middleware,
so we don't even need to think about it while writing handlers in the future.

moar for #18