fix(clock+worker): Fix bugs for the clock and worker scripts
Once again I manage to push code that won't even run. For some reason, in the tutorials I followed,
inside the request and parser callbacks 'this' was set to that object. But that isn't the case it
seems. Had to change all 'this' and 'self' to the actual request or parse object. Anyway, I've
tested the files and they run as expected.