feat(tracking): Add stat tracking and automated emails
Stat tracking is now fully functional! It also sends a very nice looking email with the stats.
Currently the system will just go through each player that's "active" and update their stats,
filtering and updating the mega stats, and then emailing the dev mail all the info. Any errors that
pop up are stored in a StatsLog collection for later debugging. There is also a bug that was patched
which caused the P2000 to be improperly tracked, because I called it 'hpk2000' instead of 'hkp2000'.
Anyways, the code is tested and seems to work just fine. Should be ready for production.

finish tracking #18