fix(clock): Fix clock interval bug introduced in 163f0ee
After adding a second job ticker to `clock.js`, and with it changing the `interval` variable, a
logic error was introduced that caused the interval value for the 'update-stats' to be used for
'check-updates' job as well. I fixed it by creating a unique interval variable for each job instead
of using a single one. Also included is a console logging update which just changes some of the
logging for the app.