fix(worker): Fixed logic error causing random articles to be emailed
When I first tested this feature, it worked perfectly. All ten of the most recent CSGO updates were
parsed, formatted, and emailed to me without problem. I got 10 unique emails, all with different
subjects and content. But now, somehow a glitch has manifested that causes the system to, no matter
what, email only the last article that is checked to me. Basically, the `item` variable was being
changed before MongoDB could insert the document, resulting in the email containing the last article
in the list. I have absolutely NO IDEA why it suddenly started doing that with 2 articles, and
worked flawlessly on the first 10. I even double checked, all 10 emails arrived with their correct
article stuff. Anyway, since the insert returns the document that was inserted, we'll now use that
as the email content instead of the changing `item` variable.