fix(android): Improve Intro.kt source code and fix UI bug
When going through the intro slides, spamming the continue button would bog down the system as it
tries to do slide changes and animations, ultimately resulting in weird UI glitches, and even the
app going completely transparent for a few seconds (lol). The source code for `Intro.kt` has been
improved with comments and better names. Although the layout of functions still makes understanding
how it works require repeately scrolling up and down the file, it is much more readable now. The
`Animer.kt` class has been improved to allow chaining of functions, and to actually work as intended
now. Added a container to the `activity_intro.xml` layout so that a safe area is now defined. This
style will be used on other layouts in the future.