feat(android): Add Player.kt and Token.kt classes
The new Player.kt class will be the way the app passes around the currently logged in player's
information. This also allows a single place to define requirements on persona, username, and email
formats. The new Token.kt class is how we'll define properties of the login token that all accounts
need to communicate with the server about private information. It holds the definition of what makes
a valid token, similar to what Player.kt does.