fix(mongodb): Fix unclosed connections bug
I thought that, maybe after like 10 hours of a connection not being used, that mongodb would just
close it. But no, it does not. In fact, the default behavoir seems to be that when you open a
connection it will NEVER CLOSE until you explicitly tell it to close, or the process restarts and
thereby killing any open connections. What the absolute fuck mongodb? Why do open connections, that
have been idle for well over 12 hours, NOT GET CLOSED?