docs(wiki): Update Wikis
Gave the wikis a much needed update. The purpose of using the repository
itself to store wiki files is because these are considered "Official"
and are the only pages that really matter. The current Wiki has many
more pages, but those are all deprecated and have been for a while. Only
now are they "officially" deprecated.

Also, I find it much easier to edit the wiki in Atom with the live
preview on the side than with GitHub. It also allows future contributors
to make change suggestions to the Wikis with pull requests.