fix(stattrak): Fix inactive-active stat tracking
Previously, if a user went from active to inactive, thereby disabling
stat tracking and updating for them, and then returned, the next update
would add everything from the inactive time period when it shouldn't.

This would result in the next time an inactive account became active, it
would think the user got all previous stat changes in a single day, and
would push them to the history and grand stats, totally breaking the
purpose of not tracking inactive accounts to skip stat changes.

Now, it checks to make sure the current NOW in the stats table is dated
in the last 2 days. If not, then the account must've gone inactive and
the worker will now just update the current stats and not save changes,
so that daily stat tracking can resume tomorrow.