fix(login): Fix login failure if Steam ID exists
While querying for Steam IDs with matching token values sounds like a
good idea, it caused the system to say that a player doesn't exist even
if they are in the Database but have an invalid token.

Now, the server looks for a matching Steam ID, saying 'not-found' only
if no result is returned, and then 'login-required' if the token is
invalid or they doesn't match. Like it should.