refactor(templates): Add max-width to email style
Although the emails look great on desktop, they do not fit on mobile,
resulting in the table going off the screen and making it impossible to
see the numbers by any method I've tried so far. So, the email now
limits how wide it can be to 100% of the viewport, but still prefers
to be 700px.