fix(database): Fix android Database.kt bug
Commas are literally my worst enemy. I also removed the not-so-important
error logging for players that aren't found in the DB, as this is
expected behavior to get an error when the player isn't found. It's up
to the caller to throw it to the logger if they want.

In case you didn't read that, there is less logger spam from Database.kt