feat(android): Add Bottom Navigation to Main.kt
Added a BottomNavigationView to the Main activity. Comes complete with
a home, history, and settings button. Included at the extra cost of 11
more lines are dynamic colors for the menu, skipping the Settings button
because in the future this will load a separate activity, so that any
changes are easily propagated back to the Main activity.