feat(android_stats): Add initial Stats-At-A-Glance
This adds all the cool stuff for the beginnings of the Stats at a Glance
feature, otherwise the Home fragment. For now it's set up to pull all
saved daily entries and only displays the aggregated time played, for
testing. Next few commits will complete the StatsAdapter.kt class and
add string resources.