refactor(tracking): Improve Stat Tracking
The clock now runs on UTC time, I didn't realize that UTC was an option
for some reason ¯\_(ツ)_/¯

Improved the email structure for stats, mainly so that it looks right
on my phone. For some reason max-width does work on the Gmail app and
I could see Mega stats at all.

The worker is in a weird place now, but it works better than before and
has more redundancy built into it. For more info, read the big TODO
comment on line 305. The feature mentioned will be added at a future
date.