fix(login): Make login update the "active" time
Stat updates look at the account's "active" time stamp to determine if
the account should be updated, however the "active" is never updated.

Just forgot to add that bit in, I guess.

I also changed the query to use the player object's Steam ID instead of
the posted Steam ID. Not sure if that matters for security, but I
thought "better safe than sorry."