refactor(strings): Add string resources for Stats
The strings used in the Home stats are now string resources, and can be
properly translated in the future. They'll still need more work done for
when we actually want to translate them, but that's detailed in the
TRANSLATING.md file.

Also, I changed the shadow radius for the actual text descriptions in
the stats. The text and small numbers have a radius of 5 while the
bigger numbers have a radius of 8.