fix(Player.kt): Fix a bug which mauls player URLs
There was a mistake that causes the URLs for the profile and avatar to
be saved as the full link, so that when the player object is recreated
it would return doubly formatted URLs.

Solved by using a private field to store the cropped url parts, and a
public getter that formats it. When saving, it saves the private field
so that future recreations don't double format the urls.