fix(login): Hide "Greetings, human!" message
There was a minor UI mistake that caused the "Greetings, human!"
message, which the API cleverly displays to crafty humans ONLY, to be
momentarily visible when logging in.

Fixed by simply hiding the WebView as soon as we know a console message
was given, so the message is shown for such a small amount of time that
I can't even tell if the text has time to be rendered at all.