style(colors): Simplify color names + BUG FIX
I simplified the way Main.kt handles the Player object. Instead of
letting 'player' be null, it is now lateinit. This means that the app
will simply crash if 'player' is ever requested before it has been set.

To fix that, there is now a 'hasPlayer' boolean variable. It is false
by default, and will only be true once a player has been logged in. I
feel this is simpler than having to constantly check if the Player is
null every single time we need some property. This new format forces us
to ensure a player is logged in at all times, and makes it easier to
know when one isn't.

This mainly simplifies the code, removing a lot of extra 'val' temporary
values and replaces with a simple 'if(hasPlayer)'.

There is also a bug fix that fails logging in, introduced by commit
9fe65ceff64c75839348dc4c896f001a354e6f98

It caused new login tokens to not be saved, skipping the initial login
phase and going straight to pulling the device token and then trying to
login with that old version, ignoring the new one it just got. This was
not noticed during testing, and I don't know why it never manifested
until yesterday when suddenly the app refused to login myself or my
tester at basically the same time.