feat(intro): Complete Intro Slide 3
The third introduction slide is FINALLY complete. It also has a really
fancy drag simulation to let the user know they can slide between two
images.

Hopefully the user will read the slide and see that it is trying to help
them streamline the login process by setting their stats to Public
before continuing. The two images together should be enough to guide the
user to the right screens to check their settings.

I limited the images to xxhdpi and xxxhdpi because I felt that smaller
versions would simply loose too much detail to be recovered with lower
available densities, and be no different from having the device manually
scale the images to fit the screen.

Animer was also expanded for use inside an ObjectAnimator listener.