fix(android): Fix internet connection bugs
Turning off the internet connection caused various error messages to
appear and crash the app. When no internet connection is available, it
now alerts the user, and continues to load whatever player information
is still on the device.

Database.kt has error message improvements for login and update player
methods, it will now always return false when failing. In the case of
no internet or api connection, it returns special identifiers to allow
the caller to construct a unique message.

This will be used in the Settings (future) and Main to display different
messages, since the context of each call will change.