docs(android): HOT DOC + Refactors
Super cool documentation for the entire app so far. I didn't document
quite everything, but all the important stuff is.

There may also be some changes to Constants.kt, InternetHelper.java, and
Database.kt. But they are minimal and don't change the functionality of
anything very much, other than making it a bit cleaner and prettier to
look at.